fix(organize-files): easier to debug under src
